Incorrect MSEX version deemed due to float conversion. Parameter changed to double.

+++ Acn/Citp/Packets/Msex/CitpMsexHeader.cs
@@ -17,7 +17,7 @@ namespace Citp.Packets
 
         #region Packet Content
 
-        public float MsexVersion { get; set; }
+        public double MsexVersion { get; set; }
 
         public string LayerContentType { get; set; }
